ai-service/requirements.txt
# ── Core API ─────────────────────────────────────────────────
fastapi==0.115.4
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
httpx==0.27.2

# ── Vector Store ─────────────────────────────────────────────
chromadb==0.5.23

# ── Embeddings (multilingual — supports Malayalam + English) ─
sentence-transformers==3.3.1

# ── Data / Validation ────────────────────────────────────────
pydantic==2.9.2
pydantic-settings==2.6.1

# ── Document Processing (Step 2: PDF/image OCR) ──────────────
PyPDF2==3.0.1
pillow==11.0.0

# ── LangChain (Step 3: Agents + LangGraph) ───────────────────
langchain==0.3.7
langchain-community==0.3.7
langchain-chroma==0.1.4
langgraph==0.2.53

# ── Google Gemini (Step 3: LLM backend) ──────────────────────
google-generativeai==0.8.3
langchain-google-genai==2.0.7
